热加载代码精简

+++ src/main/java/github/cweijan/ultimate/db/init/DBInitialer.kt
@@ -52,7 +52,8 @@ class DBInitialer(private val dbConfig: DbConfig) {
 
     }
 
-    private fun recreateTable(componentInfo: ComponentInfo) {
+    fun recreateTable(componentInfo: ComponentInfo?) {
+        if (componentInfo == null || tableExists(componentInfo.tableName)) return
         if (tableExists(componentInfo.tableName)) {
             initSqlGenetator.dropTable(componentInfo.tableName)?.let { sqlExecutor.executeSql(it) }
         }
